requirements.txt
# 核心 ASGI 框架 - 最小化配置
fastapi==0.115.13
uvicorn[standard]==0.34.3

# 數據驗證
pydantic==2.11.7

# AI 推理 - 輕量化
onnxruntime==1.22.0
numpy==2.3.0

# 圖像處理 - 最小化 (已移除 OpenCV 依賴)
pillow==11.2.1

# 網路請求
requests==2.32.4
python-multipart==0.0.20

# 速率限制 (精簡版)
slowapi==0.1.9